% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{results}
\alias{results}
\title{Fetch election results}
\usage{
results(election_date)
}
\arguments{
\item{election_date}{Election date as a string in MM-DD-YYYY format.}
}
\description{
Fetch election results
}

