% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elexr.R
\docType{package}
\name{elexr-package}
\alias{elexr}
\alias{elexr-package}
\title{Simple wrapper around the elex tool}
\description{
Wraps the elex tool to provide R access to Associated Press election results
}
\details{
A light R wrapper around the elex command

elexr provides an R interface for the elex package
(http://github.com/newsdev/elex/) to load election results from the
Associated Press Elections API.
}

